add all endpoint, one test not correct

+++ ewm-service/src/main/java/ru/practicum/category/service/CategoryServiceImpl.java
@@ -10,7 +10,9 @@ import ru.practicum.category.dto.CategoryDto;
 import ru.practicum.category.mapper.CategoryMapper;
 import ru.practicum.category.model.Category;
 import ru.practicum.category.repository.CategoryRepository;
+import ru.practicum.event.repository.EventRepository;
 import ru.practicum.exeption.ObjectNotFoundException;
+import ru.practicum.exeption.RulesViolationException;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -22,6 +24,7 @@ public class CategoryServiceImpl implements CategoryService {
     private static final Logger log = LoggerFactory.getLogger(CategoryService.class);
     private final CategoryRepository categoryRepository;
     private final CategoryMapper categoryMapper;
+    private final EventRepository eventRepository;
 
     @Override
     @Transactional
@@ -36,6 +39,9 @@ public class CategoryServiceImpl implements CategoryService {
         if (!categoryRepository.existsById(catId)) {
             throw new ObjectNotFoundException(
                     String.format("Category with id=%d was not found", catId));
+        }
+        if (eventRepository.findAllByCategoryId(catId) != null) {
+            throw new RulesViolationException(String.format("Category with id=%d can't remove, because have events", catId));
         }
         categoryRepository.deleteById(catId);
         log.info("Category was deleted");
